add hxvw script
hxvw (short for hexview) is a shell script that takes one argument which
must represent an hexadecimal color. The argument can be escaped (must
be if it starts with a sharp (#)) or can be pass as in if it's not
starting with a sharp (rrbbgg).

If the argument is correct the script sends a message to the
notification daemon with a random number content and background and
foreground colors sets to one past as arguments.

In  short, this program allows to quickly visualise an hex color from
the command line.