backup(firefox2dropbox): get firefox profile name dynamically
This patch updates the shell script that save the default firefox
profile to Dropbox. This script is triggered daily on my machine through
a cron job. The way we extract the firefox profile name is kinda rigid
atm: it parses the ini configuration file and expect a rigid structure.

At any time if the format of the configuration change, it could break
that script.