chore: archive scripts from previous laptop
This patch moves shell scripts I was using on my old laptop (with i3,
polybar and feh) to an archive directory. I might not use them in the
future as I will probably switch to sway. However I prefer to keep them
around in case of other people face the same problem or if I need a
refresher in the future.