add notify-npm-dependencies.sh script
This patch adds a new script that sends notifications through DBUS for
outdated npm packages that are globally installed. It works if several
version of node are installed (through nvm) and list all packages that
are out of date.

This patch came out before npx was getting popular and helped me to keep
track of all the dependencies I had installed globally. Nowadays this is
not as useful as before, one can use npx to use sporadically most
packages that needed to be installed globally in the past.